% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{annveronica}
\alias{annveronica}
\title{"Ann Veronica:  Modern Love Story"}
\format{A character vector with 12054 elements}
\source{
\url{https://www.gutenberg.org/files/524/524-0.txt}
}
\usage{
annveronica
}
\description{
A data set containing the complete text of H. G. Wells 1909
novel "Ann Veronica".
}
\keyword{datasets}
