% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{package}
\name{hgwellsr}
\alias{hgwellsr}
\alias{hgwellsr-package}
\alias{hgwellsr-package}
\title{Selected H. G. Wells Novels}
\description{
This package contains complete text of selected novels of
H. G. Wells formatted to be convenient for text analysis.
}
