% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{invisibleman}
\alias{invisibleman}
\title{"The Invisible Man"}
\format{A character vector with 5916 elements}
\source{
\url{https://www.gutenberg.org/files/524/524-0.txt}
}
\usage{
invisibleman
}
\description{
A data set containing the complete text of H. G. Wells 1897
novel "The Invisible Man".
}
\keyword{datasets}
