% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{mrpolly}
\alias{mrpolly}
\title{"The History of Mr. Polly"}
\format{A character vector with 8738 elements}
\source{
\url{http://www.gutenberg.org/cache/epub/7308/pg7308.txt}
}
\usage{
mrpolly
}
\description{
A data set containing the complete text of H. G. Wells 1910
novel "The History of Mr. Polly".
}
\keyword{datasets}
