% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{timemachine}
\alias{timemachine}
\title{"The Time Machine"}
\format{A character vector with 3213 elements}
\source{
\url{https://www.gutenberg.org/files/35/35-0.txt}
}
\usage{
timemachine
}
\description{
A data set containing the complete text of H. G. Wells 1895
novel "The Time Machine"
}
\keyword{datasets}
