% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{waroftheworlds}
\alias{waroftheworlds}
\title{"The War of the Worlds"}
\format{A character vector with 6456 elements}
\source{
\url{http://www.gutenberg.org/cache/epub/36/pg36.txt}
}
\usage{
waroftheworlds
}
\description{
A data set containing the complete text of H. G. Wells 1898
novel "The War of the Worlds"
}
\keyword{datasets}
